Fix a variety of CodeQL alerts

+++ src/main/java/mekanism/common/tile/machine/TileEntityIsotopicCentrifuge.java
@@ -63,7 +63,7 @@ public class TileEntityIsotopicCentrifuge extends TileEntityRecipeMachine<Chemic
           RecipeError.NOT_ENOUGH_OUTPUT_SPACE,
           RecipeError.INPUT_DOESNT_PRODUCE_OUTPUT
     );
-    public static final int MAX_GAS = 10 * FluidType.BUCKET_VOLUME;
+    public static final long MAX_GAS = 10L * FluidType.BUCKET_VOLUME;
 
     @WrappingComputerMethod(wrapper = ComputerChemicalTankWrapper.class, methodNames = {"getInput", "getInputCapacity", "getInputNeeded",
                                                                                         "getInputFilledPercentage"}, docPlaceholder = "input tank")
